microservices/upload/requirements.txt
flask
gunicorn
flask_cors
pandas
pillow
openpyxl
requests
opencv-python
numpy
matplotlib
scikit-learn
scipy
google-cloud-storage
google-cloud-vision
google-cloud-aiplatform
tqdm
aiohttp
openai
python-dotenv
ipykernel
futures
firebase-admin